Using eks addons to upgrade and downgrade vpc-cni
